Fixes #39 no private dev files and renamed where necessary
