--- app/src/main/java/com/fossasia/pslab/activity/AboutUs.java
@@ -1,20 +1,0 @@
-package com.fossasia.pslab.activity;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
-import com.fossasia.pslab.R;
-
-/**
- * Created by viveksb007 on 15/3/17.
- */
-
-public class AboutUs extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_about_us);
-    }
-}
