--- app/src/main/java/com/fossasia/pslab/activity/SplashActivity.java
@@ -1,32 +1,0 @@
-package com.fossasia.pslab.activity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
-import com.fossasia.pslab.R;
-
-/**
- * Created by viveksb007 on 11/3/17.
- */
-
-public class SplashActivity extends AppCompatActivity {
-
-    private static int SPLASH_TIME_OUT = 2000;
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.splash_screen);
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(SplashActivity.this, MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        }, SPLASH_TIME_OUT);
-    }
-}
