+++ app/src/main/java/org/fossasia/pslab/activity/AboutUs.java
@@ -1,0 +1,19 @@
+package org.fossasia.pslab.activity;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+
+import org.fossasia.pslab.R;
+/**
+ * Created by viveksb007 on 15/3/17.
+ */
+
+public class AboutUs extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_about_us);
+    }
+}
